[CRT-434] Make TaskTemplate centralised
