+++ frontend/src/components/task-instance/TaskInstanceProgressList.tsx
@@ -1,7 +1,7 @@
 import { useMemo } from "react";
 import { defineMessages, FormattedMessage, useIntl } from "react-intl";
 import styled from "@emotion/styled";
-import { Button, HStack, Icon, Status } from "@chakra-ui/react";
+import { Button, Icon } from "@chakra-ui/react";
 import { ColumnDef } from "@tanstack/react-table";
 import { useRouter } from "next/router";
 import { LuChevronRight } from "react-icons/lu";
@@ -12,7 +12,6 @@ import { CurrentStudentAnalysis } from "@/api/collimator/models/solutions/curren
 import { ColumnType } from "@/types/tanstack-types";
 import { useAllSessionTaskSolutions } from "@/api/collimator/hooks/solutions/useAllSessionTaskSolutions";
 import { useCurrentSessionTaskSolutions } from "@/api/collimator/hooks/solutions/useCurrentSessionTaskSolutions";
-import { ProgressMessages } from "@/i18n/progress-messages";
 import { EmptyState } from "@/components/EmptyState";
 import { isClickOnRow } from "@/utilities/table";
 import {
@@ -22,11 +21,7 @@ import {
 import ChakraDataTable, { ColumnSize } from "../ChakraDataTable";
 import { StudentName } from "../encryption/StudentName";
 import MultiSwrContent from "../MultiSwrContent";
-import {
-  getTaskStatus,
-  getTaskStatusColor,
-  TaskStatus,
-} from "../task-progress";
+import TaskProgress from "../task-progress";
 import StarSolutionButton from "../solution/StarSolutionButton";
 import UnstarPastSolutionsButton from "../solution/UnstarPastSolutionsButton";
 
@@ -117,50 +112,6 @@ const PreviousVersionTemplate = ({
     taskSolutions={progress.taskSolutions}
   />
 );
-
-const TaskTemplate = ({
-  classId: _classId,
-  rowData,
-}: {
-  classId: number;
-  taskId: number;
-  rowData: StudentProgress;
-}) => {
-  const intl = useIntl();
-
-  const solutionToDisplay = useMemo(
-    () => ExistingStudentSolution.findSolutionToDisplay(rowData.taskSolutions),
-    [rowData],
-  );
-
-  const status = useMemo(
-    () => getTaskStatus(solutionToDisplay, rowData.currentAnalysis),
-    [solutionToDisplay, rowData.currentAnalysis],
-  );
-
-  const color = useMemo(() => getTaskStatusColor(status), [status]);
-
-  const statusText = useMemo(() => {
-    switch (status) {
-      case TaskStatus.complete:
-        return intl.formatMessage(ProgressMessages.completeStatus);
-      case TaskStatus.incomplete:
-        return intl.formatMessage(ProgressMessages.incompleteStatus);
-      case TaskStatus.notStarted:
-      default:
-        return intl.formatMessage(ProgressMessages.notStartedStatus);
-    }
-  }, [intl, status]);
-
-  return (
-    <HStack>
-      <Status.Root>
-        <Status.Indicator backgroundColor={color} />
-      </Status.Root>
-      {statusText}
-    </HStack>
-  );
-};
 
 const TaskInstanceProgressList = ({
   classId,
@@ -265,10 +216,9 @@ const TaskInstanceProgressList = ({
         id: "progress",
         header: intl.formatMessage(messages.progressColumn),
         cell: (info) => (
-          <TaskTemplate
-            classId={classId}
-            taskId={taskId}
-            rowData={info.row.original}
+          <TaskProgress
+            solutions={info.row.original.taskSolutions}
+            currentAnalysis={info.row.original.currentAnalysis}
           />
         ),
         meta: {
